backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pgvector>=0.2.5
python-dotenv>=1.0.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-multipart>=0.0.9
